added rspec and search conditions

+++ app/controllers/api/v1/cages_controller.rb
@@ -7,17 +7,40 @@ module Api
       end
 
       def show
-        cage = Cage.find(params[:id])
-        puts "WHATS IN CAGE"
-        render json: {status: 'SUCCESS', message:'Loaded Cage ' + params[:id], data:cage.dinos}, status: :ok
+        if numeric(params[:id])
+          cage = Cage.find(params[:id])
+          render json: {status: 'SUCCESS', message:'Loaded Cage ' + params[:id], data:cage.dinos},
+          status: :ok
+          return
+        else
+          cage = Cage.where(status: params[:id])
+          render json: {status: 'SUCCESS', message:'Loaded cages of status ' + params[:id].to_s, data:cage}, status: :ok
+          return
+        end
+
       end
 
       def create
           cage = Cage.new(cage_params)
-          cage.current_capacity = 0
+          if not cage_params.include?("max_capacity")
+            render json: {status: 'ERROR', message:'Cage entry must have max_capacity variable'},
+            status: :unprocessable_entity
+            return
+          elsif not cage_params.include?("status")
+            render json: {status: 'ERROR', message:'Cage entry must have status variable'},
+            status: :unprocessable_entity
+            return
+          elsif cage.max_capacity < 0 || (cage.status != "DOWN" && cage.status != "ACTIVE")
+            render json: {status: 'ERROR', message:'Cage inputs incorrect'}, status: :unprocessable_entity
+            return
+          end
+          cage.update_attribute("current_capacity", 0)
+          cage.update_attribute("cage_type", "None")
+
 
           if cage.save
             render json: {status: 'SUCCESS', message:'saved cage', data:cage}, status: :ok
+            return
           else
             render json: {status: 'ERROR', message:'Cage not saved',
             data:cage.errors}, status: :unprocessable_entity
@@ -25,20 +48,31 @@ module Api
       end
 
       def destroy
-          cage = Cage.find(params[:id])
+
+          cageID = params[:id]
+          if not Cage.exists?(cageID)
+            render json: {status: 'ERROR', message:'Cage not found',},
+             status: :unprocessable_entity
+            return
+          end
+
+
+          cage = Cage.find(cageID)
           cage.destroy
+
+
           render json: {status: 'SUCCESS', message:'deleted cage '+ params[:id], data:cage}, status: :ok
       end
 
       def update
         cage = Cage.find(params[:id])
         if params[:status] == "DOWN" and cage.dinos.count != 0
-          render json: {status: 'ERROR', message:'Cage cannot be powered down with dinos in it'}, status: :unprocessable_entity
-
-
+          render json: {status: 'ERROR', message:'Cage cannot be powered down with dinos in it'},
+          status: :unprocessable_entity
+          return
         elsif cage.update_attributes(cage_params)
           render json: {status: 'SUCCESS', message:'updated cage ' + params[:id], data:cage}, status: :ok
-
+          return
         else
           render json: {status: 'ERROR', message:'Cage not updated',
           data:cage.errors}, status: :unprocessable_entity
@@ -46,8 +80,11 @@ module Api
       end
       private
       def cage_params
-        params.permit(:max_capacity, :current_capacity, :status, :cage_type)
+        params.permit(:max_capacity, :status, :cage_type)
 
+      end
+      def numeric(input)
+        return Float(input) != nil rescue false
       end
     end
